+++ ui/mods/bulk_create_units/keybinds.js
@@ -9,11 +9,12 @@ api.settings.definitions.keyboard.settings.bulk_paste_unit = {
   display_sub_group: 'bulk create units',
   default: 'shift+ctrl+v'
 }
-action_sets.hacks.bulk_paste_clear = function () {
-  if (model.clearPasteUnit) model.clearPasteUnit()
+
+action_sets.hacks.bulk_paste_preview_toggle = function () {
+  if (model.bulkPastePreviewToggle) model.bulkPastePreviewToggle()
 }
-api.settings.definitions.keyboard.settings.bulk_paste_clear = {
-  title: 'clear paste unit',
+api.settings.definitions.keyboard.settings.bulk_paste_preview_toggle = {
+  title: 'toggle preview',
   type: 'keybind',
   set: 'mods',
   display_group: 'mods',